Move line emission helpers to separate file, finish implementation, begin including tests
