+++ karabo/simulation/line_emission.py
@@ -2,14 +2,13 @@ import os
 import shutil
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import List, Optional, Tuple, TypeVar, Union, cast
+from typing import List, Optional, Tuple, Union, cast
 
 import h5py
 import matplotlib.pyplot as plt
 import numpy as np
 import oskar
 import xarray as xr
-from astropy.constants import c
 from astropy.convolution import Gaussian2DKernel
 from astropy.io import fits
 from astropy.wcs import WCS
@@ -17,13 +16,15 @@ from numpy.typing import NDArray
 
 from karabo.imaging.imager import Imager
 from karabo.simulation.interferometer import InterferometerSimulation, StationTypeType
+from karabo.simulation.line_emission_helpers import (
+    convert_frequency_to_z,
+    convert_z_to_frequency,
+)
 from karabo.simulation.observation import Observation
 from karabo.simulation.sky_model import SkyModel
 from karabo.simulation.telescope import Telescope
 from karabo.simulation.visibility import Visibility
 from karabo.util._types import DirPathType, FilePathType, IntFloat, NPFloatLikeStrict
-
-# from dask.delayed import Delayed
 from karabo.util.dask import DaskHandler
 from karabo.util.plotting_util import get_slices
 
@@ -239,31 +240,6 @@ def sky_slice(sky: SkyModel, z_min: IntFloat, z_max: IntFloat) -> SkyModel:
              z_max.
     """
     return sky.filter_by_column(13, z_min, z_max)
-
-
-T = TypeVar("T", NDArray[np.float_], xr.DataArray, IntFloat)
-
-
-def convert_z_to_frequency(z: T) -> T:
-    """Turn given redshift into corresponding frequency (Hz) for 21cm emission.
-
-    :param z: Redshift values to be converted into frequencies.
-
-    :return: Frequencies corresponding to input redshifts.
-    """
-
-    return cast(T, c.value / (0.21 * (1 + z)))
-
-
-def convert_frequency_to_z(freq: T) -> T:
-    """Turn given frequency (Hz) into corresponding redshift for 21cm emission.
-
-    :param freq: Frequency values to be converted into redshifts.
-
-    :return: Redshifts corresponding to input frequencies.
-    """
-
-    return cast(T, (c.value / (0.21 * freq)) - 1)
 
 
 def freq_channels(
