+++ karabo/simulation/line_emission_helpers.py
@@ -1,0 +1,32 @@
+from typing import TypeVar, cast
+
+import numpy as np
+import xarray as xr
+from astropy.constants import c
+from numpy.typing import NDArray
+
+from karabo.util._types import IntFloat
+
+T = TypeVar("T", NDArray[np.float_], xr.DataArray, IntFloat)
+
+
+def convert_z_to_frequency(z: T) -> T:
+    """Turn given redshift into corresponding frequency (Hz) for 21cm emission.
+
+    :param z: Redshift values to be converted into frequencies.
+
+    :return: Frequencies corresponding to input redshifts.
+    """
+
+    return cast(T, c.value / (0.21 * (1 + z)))
+
+
+def convert_frequency_to_z(freq: T) -> T:
+    """Turn given frequency (Hz) into corresponding redshift for 21cm emission.
+
+    :param freq: Frequency values to be converted into redshifts.
+
+    :return: Redshifts corresponding to input frequencies.
+    """
+
+    return cast(T, (c.value / (0.21 * freq)) - 1)
