+++ karabo/test/test_skymodel.py
@@ -146,3 +146,7 @@ def test_read_healpix_map():
 
 def test_get_poisson_sky():
     _ = SkyModel.get_random_poisson_disk_sky((220, -60), (260, -80), 0.1, 0.8, 2)
+
+
+def test_convert_sky_to_RASCIL():
+    pass
